# doc/api-doc.yml
openapi: 3.0.0
info:
  title: 武大论坛接口文档
  version: 1.0.0
servers:
  - url: whuerbbs.cn/v1
paths:
...
  /posts/{post_id}:
    get:
      summary: 获取帖子详情
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostVO'
...
  /secondhand_posts:
    get:
      summary: 获取全部二手交易帖子
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PostListVO'
                  - type: object
                    properties:
                      trade_category:
                        $ref: '#/components/schemas/TradeCategory'
                      campus:
                        $ref: '#/components/schemas/Campus'
    post:
      summary: 发布二手交易帖子
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/PostDTO'
                - type: object
                  properties:
                    trade_category:
                      $ref: '#/components/schemas/TradeCategory'
                    campus:
                      $ref: '#/components/schemas/Campus'
      responses:
        200:
          $ref: '#/components/responses/Ok'
  /secondhand_posts/{secondhand_post_id}:
    get:
      summary: 获取二手交易帖子详情
      parameters:
        - in: path
          name: secondhand_post_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PostVO'
                  - type: object
                    properties:
                      trade_category:
                        $ref: '#/components/schemas/TradeCategory'
                      campus:
                        $ref: '#/components/schemas/Campus'
...
components:
  responses:
    Ok:
      description: 成功
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
              message:
                type: integer
...
  schemas:
    AttitudeStatus:
      type: integer
      enum: [0, 1]
      description: >
        Sort order:
         * `0` - 默认
         * `1` - 点赞
...
    Board:
      type: integer
      enum: [1, 2, 3, 4, 5]
      description: >
        Sort order:
          * `1` - 新鲜事
          * `2` - 二手区
          * `3` - 问点事
          * `4` - 匿名墙
          * `5` - 一起玩
...
    UserVO:
      type: object
      properties:
        id:
          type: string
        nickname:
          type: string
        avatar_url:
          type: string
...
    TopicVO:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
    PostDTO:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        images:
          type: array
          items:
            type: string
    PostListVO:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        summary:
          type: string
        board:
          $ref: '#/components/schemas/Board'
        image:
          type: string
        like_count:
          type: integer
        comment_count:
          type: integer
        created_at:
          type: string
          format: 'date-time'
        last_active_at:
          type: string
          format: 'date-time'
        created_by:
          $ref: '#/components/schemas/UserVO'
        topics:
          type: array
          items:
            $ref: '#/components/schemas/TopicVO'
    PostVO:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        board:
          $ref: '#/components/schemas/Board'
        images:
          type: array
          items:
            type: string
        like_count:
          type: integer
        comment_count:
          type: integer
        last_active_at:
          type: integer
        attitude_status:
          $ref: '#/components/schemas/AttitudeStatus'
        created_by:
          $ref: '#/components/schemas/UserVO'
        topics:
          type: array
          items:
            $ref: '#/components/schemas/TopicVO'
        hot:
          type: boolean
        collected:
          type: boolean
        can_delete:
          type: boolean
...
    TradeCategory:
      type: integer
      enum: [1, 2, 3, 4, 5]
      description: >
        Sort order:
         * `1` - 综合
         * `2` - 数码
         * `3` - 衣物
         * `4` - 书刊
         * `5` - 生活
    Campus:
      type: integer
      enum: [1, 2, 3, 4, 5, 6]
      description: >
        Sort order:
         * `1` - 全校区
         * `2` - 文理学部
         * `3` - 信息学部
         * `4` - 工学部
         * `5` - 医学部
         * `6` - 线上
...